+++ servant-multipart-api/servant-multipart-api.cabal
@@ -33,8 +33,7 @@ library
 
   -- other dependencies
   build-depends:
-      resourcet           >=1.2.2    && <1.3
-    , servant             >=0.16     && <0.19
+      servant             >=0.16     && <0.19
 
 source-repository head
   type:     git
