+++ servant-multipart-client/servant-multipart-client.cabal
@@ -45,13 +45,22 @@ executable upload
   default-language: Haskell2010
   build-depends:
       base
-    , bytestring
     , http-client
     , network            >=2.8 && <3.2
     , servant
-    , servant-multipart
+    , servant-multipart-api
     , servant-multipart-client
     , servant-client
     , servant-client-core
+
+executable server
+  hs-source-dirs:   exe
+  main-is:          Server.hs
+  default-language: Haskell2010
+  build-depends:
+      base
+    , bytestring
+    , network            >=2.8 && <3.2
+    , servant-multipart
     , servant-server
     , warp
